Comeback of the mplayer backend
It once used to be the default and only backend. It has been replaced by
gstreamer to ease packaging and clean the architecture (single process).
I brought it back since I use older machines, and I have been
maintaining a separate build in a separate directory for quite some
time. It's a compromise, and it's here as a reminder of the cruel
reality of backward compatibility.
It might not stay forever, since I still believe the gstreamer solution
is cleaner.